chore: [app-router-migration 26] Migrate "/[user]" page (#13057)
