+++ apps/web/pages/[user]/[type]/index.tsx
@@ -1,12 +1,10 @@
 import { type AppProps } from "@lib/app-providers";
-import withEmbedSsr from "@lib/withEmbedSsr";
 
 import PageWrapper, { type CalPageWrapper } from "@components/PageWrapper";
 
 import TypePage from "~/users/views/users-type-public-view";
-import { getServerSideProps as _getServerSideProps } from "~/users/views/users-type-public-view.getServerSideProps";
 
-export const getServerSideProps = withEmbedSsr(_getServerSideProps);
+export { getServerSideProps } from "~/users/views/users-type-public-view.getServerSideProps";
 
 const Type = TypePage as unknown as CalPageWrapper & {
   isBookingPage: AppProps["Component"]["isBookingPage"];
