+++ apps/web/pages/org/[orgSlug]/[user]/[type]/embed.tsx
@@ -2,8 +2,8 @@
 
 import withEmbedSsr from "@lib/withEmbedSsr";
 
-import { getServerSideProps as _getServerSideProps } from "../[type]";
+import { getServerSideProps as _getServerSideProps } from ".";
 
-export { default, type PageProps } from "../[type]";
+export { default, type PageProps } from ".";
 
 export const getServerSideProps = withEmbedSsr(_getServerSideProps);
