+++ apps/web/pages/org/[orgSlug]/[user]/[type]/index.tsx
@@ -2,10 +2,11 @@
 
 import PageWrapper from "@components/PageWrapper";
 
-import type { PageProps as UserTypePageProps } from "../../../[user]/[type]";
-import UserTypePage from "../../../[user]/[type]";
-import type { PageProps as TeamTypePageProps } from "../../../team/[slug]/[type]";
-import TeamTypePage from "../../../team/[slug]/[type]";
+import UserTypePage from "~/users/views/users-type-public-view";
+import type { PageProps as UserTypePageProps } from "~/users/views/users-type-public-view.getServerSideProps";
+
+import type { PageProps as TeamTypePageProps } from "../../../../team/[slug]/[type]";
+import TeamTypePage from "../../../../team/[slug]/[type]";
 
 export { getServerSideProps } from "@lib/org/[orgSlug]/[user]/[type]/getServerSideProps";
